+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# [0.2.0] - 2023-10-30
+# Unreleased - 2023-10-30
+
+### Added
+- Capture command
+- CaptureFit command 
+- Extra help for commands (`help command_name`)
+
+# 0.2.0 - 2023-10-30
 
 ### Added
 - Command input below log panel
@@ -11,7 +18,7 @@
 ### Changed
 - Bump click_extra to 4.7.1
 
-# [O.1.0-beta] - 2023-10-26
+# O.1.0-beta - 2023-10-26
 
 ### Added